Show card states
The card component should reflect the state of the card and display accordingly. We can use the new built in control flow to switch the UI based on the card state

I added a temporary state switcher method on click for demonstrative purposes. This will work a little differently once the game is implemented since it needs to trigger the game to run.