Flip card implementation
The user should be able to flip two cards and compare them to determine if there is a match. For this I utilized TDD to describe the logic needed and then implemented it accordingly.

I decided to split this into two methods in the end so that the user can click anywhere to run the comparison after both cards are revealed